rename component to TaxReliefDataProvider
