+++ webapp/app/layout.tsx
@@ -6,7 +6,7 @@ import "@newjersey/njwds/dist/css/styles.css";
 import "./globals.css";
 import Script from "next/script";
 import { FeedbackWidget } from "@/components/FeedbackWidget";
-import { TaxReliefStatusProvider } from "./dataStore";
+import { TaxReliefDataProvider } from "../components/TaxReliefDataProvider";
 
 /** {@link https://nextjs.org/docs/app/api-reference/functions/generate-metadata} */
 export const metadata: Metadata = {
@@ -36,13 +36,13 @@ const RootLayout = ({ children }: { readonly children: React.ReactNode }) => (
       )}
     </head>
     <body>
-      <TaxReliefStatusProvider>
+      <TaxReliefDataProvider>
         <NjHeader />
         <StatusCheckerHeader />
         {children}
         <FeedbackWidget />
         <NjFooter />
-      </TaxReliefStatusProvider>
+      </TaxReliefDataProvider>
     </body>
   </html>
 );
