+++ webapp/app/status/page.tsx
@@ -3,7 +3,7 @@
 import { useEffect } from "react";
 import { useRouter } from "next/navigation";
 import Link from "next/link";
-import { useDataStore } from "../dataStore";
+import { useDataStore } from "../../components/TaxReliefDataProvider";
 
 const StatusPage = () => {
   const router = useRouter();
